+++ src/pages/JobTitle/JobTitleWorkExperiencesProvider.js
@@ -21,7 +21,7 @@ import {
 } from 'selectors/routing';
 import { getFetched, isFetched } from 'utils/fetchBox';
 
-import useJobTitle, { jobTitleSelector } from './useJobTitle';
+import useJobTitleParam, { jobTitleSelector } from './useJobTitleParam';
 
 const useWorkExperiencesBoxSelector = pageName => {
   return useCallback(
@@ -47,7 +47,7 @@ const useWorkExperiencesBoxSelector = pageName => {
 const JobTitleWorkExperiencesProvider = () => {
   const dispatch = useDispatch();
   const pageType = PageType.JOB_TITLE;
-  const jobTitle = useJobTitle();
+  const jobTitle = useJobTitleParam();
   const [companyName] = useSearchTextFromQuery();
   const [sortBy] = useSortByFromQuery();
   const page = usePage();
