+++ src/pages/JobTitle/useJobTitleParam.ts
@@ -6,9 +6,10 @@ type Params = { jobTitle: string };
 export const jobTitleSelector = (params: Params): string =>
   decodeURIComponent(params.jobTitle);
 
-const useJobTitle = (): string => {
+// 名字帶 Param 的理由見 pages/Company/useCompanyNameParam
+const useJobTitleParam = (): string => {
   const params = useParams<Params>();
   return jobTitleSelector(params);
 };
 
-export default useJobTitle;
+export default useJobTitleParam;
